refactor: split Query and normal controller
