+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/ProcessInstanceController.java
@@ -9,25 +9,19 @@ package io.camunda.zeebe.gateway.rest.controller;
 
 import io.camunda.service.ProcessInstanceServices;
 import io.camunda.service.ProcessInstanceServices.ProcessInstanceStartRequest;
-import io.camunda.service.search.query.ProcessInstanceQuery;
-import io.camunda.zeebe.gateway.protocol.rest.ProcessInstanceSearchQueryRequest;
-import io.camunda.zeebe.gateway.protocol.rest.ProcessInstanceSearchQueryResponse;
 import io.camunda.zeebe.gateway.protocol.rest.StartProcessInstanceRequest;
 import io.camunda.zeebe.gateway.rest.RequestMapper;
 import io.camunda.zeebe.gateway.rest.ResponseMapper;
 import io.camunda.zeebe.gateway.rest.RestErrorMapper;
-import io.camunda.zeebe.gateway.rest.SearchQueryRequestMapper;
-import io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper;
 import java.util.concurrent.CompletableFuture;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-@CamundaRestQueryController
+@CamundaRestController
 @RequestMapping("/v2/process-instances")
 public class ProcessInstanceController {
 
@@ -40,35 +34,6 @@ public class ProcessInstanceController {
       @RequestBody final StartProcessInstanceRequest request) {
     return RequestMapper.toStartProcessInstance(request)
         .fold(RestErrorMapper::mapProblemToCompletedResponse, this::startProcessInstance);
-  }
-
-  @PostMapping(
-      path = "/search",
-      produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_PROBLEM_JSON_VALUE},
-      consumes = MediaType.APPLICATION_JSON_VALUE)
-  public ResponseEntity<ProcessInstanceSearchQueryResponse> searchProcessInstances(
-      @RequestBody(required = false) final ProcessInstanceSearchQueryRequest query) {
-    return SearchQueryRequestMapper.toProcessInstanceQuery(query)
-        .fold(RestErrorMapper::mapProblemToResponse, this::search);
-  }
-
-  private ResponseEntity<ProcessInstanceSearchQueryResponse> search(
-      final ProcessInstanceQuery query) {
-    try {
-      final var result =
-          processInstanceServices
-              .withAuthentication(RequestMapper.getAuthentication())
-              .search(query);
-      return ResponseEntity.ok(
-          SearchQueryResponseMapper.toProcessInstanceSearchQueryResponse(result));
-    } catch (final Throwable e) {
-      final var problemDetail =
-          RestErrorMapper.createProblemDetail(
-              HttpStatus.BAD_REQUEST,
-              e.getMessage(),
-              "Failed to execute Process Instance Search Query");
-      return RestErrorMapper.mapProblemToResponse(problemDetail);
-    }
   }
 
   private CompletableFuture<ResponseEntity<Object>> startProcessInstance(
