+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/configuration/RestApiQueryEnabledTest.java
@@ -11,13 +11,13 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 
 import io.camunda.service.search.query.ProcessInstanceQuery;
-import io.camunda.zeebe.gateway.rest.controller.ProcessInstanceController;
+import io.camunda.zeebe.gateway.rest.controller.ProcessInstanceQueryController;
 import io.camunda.zeebe.gateway.rest.controller.TopologyController;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 
 @WebMvcTest(
-    value = {ProcessInstanceController.class, TopologyController.class},
+    value = {ProcessInstanceQueryController.class, TopologyController.class},
     properties = "camunda.rest.query.enabled=true")
 public class RestApiQueryEnabledTest extends RestApiConfigurationTest {
