+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/ProcessInstanceControllerTest.java
@@ -9,22 +9,15 @@ package io.camunda.zeebe.gateway.rest.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import io.camunda.service.ProcessInstanceServices;
 import io.camunda.service.ProcessInstanceServices.ProcessInstanceStartRequest;
-import io.camunda.service.entities.ProcessInstanceEntity;
-import io.camunda.service.search.query.ProcessInstanceQuery;
-import io.camunda.service.search.query.SearchQueryResult;
-import io.camunda.service.search.query.SearchQueryResult.Builder;
-import io.camunda.service.search.sort.ProcessInstanceSort;
 import io.camunda.service.security.auth.Authentication;
 import io.camunda.zeebe.gateway.rest.RestControllerTest;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceCreationRecord;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceResultRecord;
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -34,38 +27,8 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 
-@WebMvcTest(value = ProcessInstanceController.class, properties = "camunda.rest.query.enabled=true")
+@WebMvcTest(value = ProcessInstanceController.class)
 public class ProcessInstanceControllerTest extends RestControllerTest {
-
-  static final String EXPECTED_SEARCH_RESPONSE =
-      """
-      {
-          "items": [
-            {
-              "key": 123,
-              "processName": "Demo Process",
-              "processVersion": 5,
-              "bpmnProcessId": "demoProcess",
-              "parentKey": 555,
-              "parentFlowNodeInstanceKey": 789,
-              "startDate": "2024-01-01T00:00:00Z",
-              "state": "ACTIVE",
-              "incident": false,
-              "hasActiveOperation": false,
-              "processDefinitionKey": 777,
-              "tenantId": "default"
-            }
-          ],
-          "page": {
-              "totalItems": 1,
-              "firstSortValues": [],
-              "lastSortValues": [
-                  "v"
-              ]
-          }
-      }""";
-  private static final String PROCESS_INSTANCES_SEARCH_URL = "/v2/process-instances/search";
-  private static final SearchQueryResult<ProcessInstanceEntity> SEARCH_QUERY_RESULT =
 
   static final String EXPECTED_START_RESPONSE =
       """
@@ -76,297 +39,16 @@ public class ProcessInstanceControllerTest extends RestControllerTest {
          "processInstanceKey":123,
          "tenantId":"tenantId"
       }""";
-
-  static final SearchQueryResult<ProcessInstanceEntity> SEARCH_QUERY_RESULT =
-      new Builder<ProcessInstanceEntity>()
-          .total(1L)
-          .items(
-              List.of(
-                  new ProcessInstanceEntity(
-                      123L,
-                      "Demo Process",
-                      5,
-                      "demoProcess",
-                      555L,
-                      789L,
-                      "2024-01-01T00:00:00Z",
-                      null,
-                      "ACTIVE",
-                      false,
-                      false,
-                      777L,
-                      "default",
-                      null,
-                      null,
-                      null)))
-          .sortValues(new Object[] {"v"})
-          .build();
-
   static final String PROCESS_INSTANCES_START_URL = "/v2/process-instances";
 
-  @MockBean ProcessInstanceServices processInstanceServices;
   @Captor ArgumentCaptor<ProcessInstanceStartRequest> requestCaptor;
+
+  @MockBean ProcessInstanceServices processInstanceServices;
 
   @BeforeEach
   void setupServices() {
     when(processInstanceServices.withAuthentication(any(Authentication.class)))
         .thenReturn(processInstanceServices);
-  }
-
-  @Test
-  void shouldSearchProcessInstancesWithEmptyBody() {
-    // given
-    when(processInstanceServices.search(any(ProcessInstanceQuery.class)))
-        .thenReturn(SEARCH_QUERY_RESULT);
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_JSON)
-        .expectBody()
-        .json(EXPECTED_SEARCH_RESPONSE);
-
-    verify(processInstanceServices).search(new ProcessInstanceQuery.Builder().build());
-  }
-
-  @Test
-  void shouldSearchProcessInstancesWithEmptyQuery() {
-    // given
-    when(processInstanceServices.search(any(ProcessInstanceQuery.class)))
-        .thenReturn(SEARCH_QUERY_RESULT);
-    final String request = "{}";
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_JSON)
-        .expectBody()
-        .json(EXPECTED_SEARCH_RESPONSE);
-
-    verify(processInstanceServices).search(new ProcessInstanceQuery.Builder().build());
-  }
-
-  @Test
-  void shouldSearchProcessInstancessWithSorting() {
-    // given
-    when(processInstanceServices.search(any(ProcessInstanceQuery.class)))
-        .thenReturn(SEARCH_QUERY_RESULT);
-    final var request =
-        """
-        {
-            "sort": [
-                {
-                    "field": "bpmnProcessId",
-                    "order": "desc"
-                },
-                {
-                    "field": "processDefinitionKey",
-                    "order": "asc"
-                }
-            ]
-        }""";
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isOk()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_JSON)
-        .expectBody()
-        .json(EXPECTED_SEARCH_RESPONSE);
-
-    verify(processInstanceServices)
-        .search(
-            new ProcessInstanceQuery.Builder()
-                .sort(
-                    new ProcessInstanceSort.Builder()
-                        .bpmnProcessId()
-                        .desc()
-                        .processDefinitionKey()
-                        .asc()
-                        .build())
-                .build());
-  }
-
-  @Test
-  void shouldInvalidateProcessInstancesSearchQueryWithBadSortOrder() {
-    // given
-    final var request =
-        """
-        {
-            "sort": [
-                {
-                    "field": "bpmnProcessId",
-                    "order": "dsc"
-                }
-            ]
-        }""";
-    final var expectedResponse =
-        String.format(
-            """
-        {
-          "type": "about:blank",
-          "title": "INVALID_ARGUMENT",
-          "status": 400,
-          "detail": "Unknown sortOrder: dsc.",
-          "instance": "%s"
-        }""",
-            PROCESS_INSTANCES_SEARCH_URL);
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isBadRequest()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_PROBLEM_JSON)
-        .expectBody()
-        .json(expectedResponse);
-
-    verify(processInstanceServices, never()).search(any(ProcessInstanceQuery.class));
-  }
-
-  @Test
-  void shouldInvalidateProcessInstancesSearchQueryWithBadSortField() {
-    // given
-    final var request =
-        """
-        {
-            "sort": [
-                {
-                    "field": "unknownField",
-                    "order": "asc"
-                }
-            ]
-        }""";
-    final var expectedResponse =
-        String.format(
-            """
-        {
-          "type": "about:blank",
-          "title": "INVALID_ARGUMENT",
-          "status": 400,
-          "detail": "Unknown sortBy: unknownField.",
-          "instance": "%s"
-        }""",
-            PROCESS_INSTANCES_SEARCH_URL);
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isBadRequest()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_PROBLEM_JSON)
-        .expectBody()
-        .json(expectedResponse);
-
-    verify(processInstanceServices, never()).search(any(ProcessInstanceQuery.class));
-  }
-
-  @Test
-  void shouldInvalidateProcessInstancesSearchQueryWithMissingSortField() {
-    // given
-    final var request =
-        """
-        {
-            "sort": [
-                {
-                    "order": "asc"
-                }
-            ]
-        }""";
-    final var expectedResponse =
-        String.format(
-            """
-        {
-          "type": "about:blank",
-          "title": "INVALID_ARGUMENT",
-          "status": 400,
-          "detail": "Sort field must not be null.",
-          "instance": "%s"
-        }""",
-            PROCESS_INSTANCES_SEARCH_URL);
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isBadRequest()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_PROBLEM_JSON)
-        .expectBody()
-        .json(expectedResponse);
-
-    verify(processInstanceServices, never()).search(any(ProcessInstanceQuery.class));
-  }
-
-  @Test
-  void shouldInvalidateProcessInstancesSearchQueryWithConflictingPagination() {
-    // given
-    final var request =
-        """
-        {
-            "page": {
-                "searchAfter": ["a"],
-                "searchBefore": ["b"]
-            }
-        }""";
-    final var expectedResponse =
-        String.format(
-            """
-        {
-          "type": "about:blank",
-          "title": "INVALID_ARGUMENT",
-          "status": 400,
-          "detail": "Both searchAfter and searchBefore cannot be set at the same time.",
-          "instance": "%s"
-        }""",
-            PROCESS_INSTANCES_SEARCH_URL);
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_SEARCH_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isBadRequest()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_PROBLEM_JSON)
-        .expectBody()
-        .json(expectedResponse);
-
-    verify(processInstanceServices, never()).search(any(ProcessInstanceQuery.class));
   }
 
   @Test
@@ -587,43 +269,7 @@ public class ProcessInstanceControllerTest extends RestControllerTest {
             "type":"about:blank",
             "title":"INVALID_ARGUMENT",
             "status":400,
-            "detail":"Only one of [bpmnProcessId, processDefinitionKey] is required.",
-            "instance":"/v2/process-instances"
-         }""";
-
-    // when / then
-    webClient
-        .post()
-        .uri(PROCESS_INSTANCES_START_URL)
-        .accept(MediaType.APPLICATION_JSON)
-        .contentType(MediaType.APPLICATION_JSON)
-        .bodyValue(request)
-        .exchange()
-        .expectStatus()
-        .isBadRequest()
-        .expectHeader()
-        .contentType(MediaType.APPLICATION_PROBLEM_JSON)
-        .expectBody()
-        .json(expectedBody);
-  }
-
-  @Test
-  void shouldRejectStartProcessInstancesIfBpmnProcessIdIsProvidedButNoVersion() {
-    // given
-    final var request =
-        """
-        {
-            "bpmnProcessId": "bpmnProcessId",
-            "awaitCompletion": true
-        }""";
-
-    final var expectedBody =
-        """
-        {
-            "type":"about:blank",
-            "title":"INVALID_ARGUMENT",
-            "status":400,
-            "detail":"If bpmnProcessId is provided, version is required.",
+            "detail":"Only one of [bpmnProcessId, processDefinitionKey] is allowed.",
             "instance":"/v2/process-instances"
          }""";
